[TASK] Move indexQueueAction to IndexQueueModuleController
With this commit clearing the index queue works again.

Closes: #3087
