+++ Classes/Controller/Backend/Search/IndexAdministrationModuleController.php
@@ -28,11 +28,8 @@ use Psr\Http\Message\ResponseInterface;
 use ApacheSolrForTypo3\Solr\ConnectionManager;
 use ApacheSolrForTypo3\Solr\IndexQueue\Queue;
 use ApacheSolrForTypo3\Solr\System\Solr\SolrConnection;
-use ApacheSolrForTypo3\Solr\Util;
-use TYPO3\CMS\Backend\Routing\UriBuilder as BackendUriBuilder;
 use TYPO3\CMS\Core\Messaging\FlashMessage;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use TYPO3\CMS\Extbase\Mvc\Web\ReferringRequest;
 use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
 
 /**
@@ -113,22 +110,6 @@ class IndexAdministrationModuleController extends AbstractModuleController
     }
 
     /**
-     * Empties the Index Queue
-     *
-     * @return void
-     */
-    public function clearIndexQueueAction(): ResponseInterface
-    {
-        $this->indexQueue->deleteItemsBySite($this->selectedSite);
-        $this->addFlashMessage(
-            LocalizationUtility::translate('solr.backend.index_administration.success.queue_emptied', 'Solr',
-                [$this->selectedSite->getLabel()])
-        );
-        $this->redirectToReferrerModule();
-        return $this->htmlResponse();
-    }
-
-    /**
      * Reloads the site's Solr cores.
      *
      * @return void
@@ -168,31 +149,5 @@ class IndexAdministrationModuleController extends AbstractModuleController
         }
 
         $this->redirect('index');
-    }
-
-    /**
-     * Redirects to the referrer module index Action.
-     *
-     * Fluids <f:form VH can not make urls to other modules properly.
-     * The module name/key is not provided in the hidden fields __referrer by bulding form.
-     * So this is currently the single way to make it possible.
-     *
-     * @todo: remove this method if f:form works properly between backend modules.
-     */
-    protected function redirectToReferrerModule()
-    {
-        $wasFromQueue = $this->request->hasArgument('fromQueue');
-        if (!$wasFromQueue) {
-            $this->redirect('index');
-            return;
-        }
-
-        /* @var BackendUriBuilder $backendUriBuilder */
-        $backendUriBuilder = GeneralUtility::makeInstance(BackendUriBuilder::class);
-
-        $parameters =  ['id' => $this->selectedPageUID];
-        $referringUri = $backendUriBuilder->buildUriFromRoute('searchbackend_SolrIndexqueue', $parameters);
-
-        $this->redirectToUri($referringUri);
     }
 }
