+++ Resources/Private/Templates/Backend/Search/IndexQueueModule/Index.html
@@ -128,9 +128,7 @@
 
 			<div class="row section-with-header">
 				<div class="col-md-3">
-
-					<f:form addQueryString="1"
-							additionalParams="{route: '/searchbackend/SolrIndexadministration/', tx_solr_searchbackend_solrindexadministration:{controller: 'Backend\\Search\\IndexAdministrationModule', action: 'clearIndexQueue', id: '{pageUID}', fromQueue: 'true'}}">
+					<f:form action="clearIndexQueue" additionalParams="{id: '{pageUID}'}">
 						<f:form.submit class="btn btn-sm btn-default btn-danger t3js-modal-formsubmit-trigger"
 									   data="{title: 'Please confirm', content: 'Are you sure you want to clear the Index Queue?', severity: 'warning'}"
 									   value="{f:translate(key:'solr.backend.index_queue_module.button.clear_index_queue')}"/>
